Alan:  fix up tests for changing mat to int

+++ test/junit/abfab3d/grid/BaseTestGrid.java
@@ -55,7 +55,7 @@ public class BaseTestGrid extends TestCase {
             }
         }
     }
-    
+
     /**
      * Set and get all values of a grid using world coords
      *
@@ -69,24 +69,24 @@ public class BaseTestGrid extends TestCase {
         double sliceHeight = grid.getSliceHeight();
 
         double xcoord, ycoord, zcoord;
-        
+
         for(int x=0; x < width; x++) {
-        	xcoord = (double)(x)*voxelSize + voxelSize/2.0;
+            xcoord = (double)(x)*voxelSize + voxelSize/2.0;
             for(int y=0; y < height; y++) {
-            	ycoord = (double)(y)*sliceHeight + sliceHeight/2.0;
+                ycoord = (double)(y)*sliceHeight + sliceHeight/2.0;
                 for(int z=0; z < depth; z++) {
-                	zcoord = (double)(z)*voxelSize + voxelSize/2.0;
+                    zcoord = (double)(z)*voxelSize + voxelSize/2.0;
                     grid.setData(xcoord, ycoord, zcoord, Grid.EXTERIOR, (byte)1);
                 }
             }
         }
 
         for(int x=0; x < width; x++) {
-        	xcoord = (double)(x)*voxelSize + voxelSize/2.0;
+            xcoord = (double)(x)*voxelSize + voxelSize/2.0;
             for(int y=0; y < height; y++) {
-            	ycoord = (double)(y)*sliceHeight + sliceHeight/2.0;
+                ycoord = (double)(y)*sliceHeight + sliceHeight/2.0;
                 for(int z=0; z < depth; z++) {
-                	zcoord = (double)(z)*voxelSize + voxelSize/2.0;
+                    zcoord = (double)(z)*voxelSize + voxelSize/2.0;
                     VoxelData vd = grid.getData(xcoord, ycoord, zcoord);
 //System.out.println(x + ", " + y + ", " + z + ": " + vd.getState());
                     assertTrue("State wrong", vd.getState() == Grid.EXTERIOR);
@@ -95,7 +95,7 @@ public class BaseTestGrid extends TestCase {
             }
         }
     }
-    
+
     /**
      * Set the X values of a grid with a given Y and Z to the given state and material.
      *
@@ -104,12 +104,12 @@ public class BaseTestGrid extends TestCase {
      * @param startIndex The starting X index
      * @param endIndex The ending X Index
      */
-    protected static void setX(Grid grid, int y, int z, byte state, byte mat, int startIndex, int endIndex) {
+    protected static void setX(Grid grid, int y, int z, byte state, int mat, int startIndex, int endIndex) {
         for(int x=startIndex; x <= endIndex; x++) {
             grid.setData(x,y,z, state, mat);
         }
     }
-    
+
     /**
      * Set the Y values of a grid with a given X and Z to the given state and material.
      *
@@ -118,12 +118,12 @@ public class BaseTestGrid extends TestCase {
      * @param startIndex The starting Y index
      * @param endIndex The ending Y Index
      */
-    protected static void setY(Grid grid, int x, int z, byte state, byte mat, int startIndex, int endIndex) {
+    protected static void setY(Grid grid, int x, int z, byte state, int mat, int startIndex, int endIndex) {
         for(int y=startIndex; y <= endIndex; y++) {
             grid.setData(x,y,z, state, mat);
         }
     }
-    
+
     /**
      * Set the Z values of a grid with a given X and Y to the given state and material.
      *
@@ -132,13 +132,13 @@ public class BaseTestGrid extends TestCase {
      * @param startIndex The starting Z index
      * @param endIndex The ending Z Index
      */
-    protected static void setZ(Grid grid, int x, int y, byte state, byte mat, int startIndex, int endIndex) {
+    protected static void setZ(Grid grid, int x, int y, byte state, int mat, int startIndex, int endIndex) {
         for(int z=startIndex; z <= endIndex; z++) {
             grid.setData(x,y,z, state, mat);
         }
     }
-    
-    
+
+
     /**
      * Set the data for an X plane.
      *
@@ -148,16 +148,16 @@ public class BaseTestGrid extends TestCase {
      * @param mat The new material
      */
     protected static void setPlaneX(Grid grid, int x, byte state, int material) {
-    	int height = grid.getHeight();
-    	int depth = grid.getDepth();
-    	
-    	for (int y=0; y<height; y++) {
-    		for (int z=0; z<depth; z++) {
-    			grid.setData(x, y, z, state, material);
-    		}
-    	}
+        int height = grid.getHeight();
+        int depth = grid.getDepth();
+
+        for (int y=0; y<height; y++) {
+            for (int z=0; z<depth; z++) {
+                grid.setData(x, y, z, state, material);
+            }
+        }
     }
-    
+
     /**
      * Set the data for a Y plane.
      *
@@ -167,16 +167,16 @@ public class BaseTestGrid extends TestCase {
      * @param mat The new material
      */
     protected static void setPlaneY(Grid grid, int y, byte state, int material) {
-    	int width = grid.getWidth();
-    	int depth = grid.getDepth();
-    	
-    	for (int x=0; x<width; x++) {
-    		for (int z=0; z<depth; z++) {
-    			grid.setData(x, y, z, state, material);
-    		}
-    	}
+        int width = grid.getWidth();
+        int depth = grid.getDepth();
+
+        for (int x=0; x<width; x++) {
+            for (int z=0; z<depth; z++) {
+                grid.setData(x, y, z, state, material);
+            }
+        }
     }
-    
+
     /**
      * Set the data for a Z plane.
      *
@@ -186,13 +186,13 @@ public class BaseTestGrid extends TestCase {
      * @param mat The new material
      */
     protected static void setPlaneZ(Grid grid, int z, byte state, int material) {
-    	int width = grid.getWidth();
-    	int height = grid.getHeight();
-    	
-    	for (int x=0; x<width; x++) {
-    		for (int y=0; y<height; y++) {
-    			grid.setData(x, y, z, state, material);
-    		}
-    	}
+        int width = grid.getWidth();
+        int height = grid.getHeight();
+
+        for (int x=0; x<width; x++) {
+            for (int y=0; y<height; y++) {
+                grid.setData(x, y, z, state, material);
+            }
+        }
     }
 }
